requirements.txt
streamlit
python-dotenv
requests
groq
orjson
